requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==0.19.0
flask-cors==3.0.10 